requirements.txt
numpy==2.2.5
pygame-ce==2.5.3
